18 y 19

+++ Ejer18/aritmetica.c
@@ -11,6 +11,6 @@ int resta  (int minuendo, int sustraendo) {
 int multiplicacion (int  numero1, int numero2) {
 	return (numero1*numero2);
 
-int division (int dividendo, int divisor) {
-	return (dividendo/divisor);
+float division (int dividendo, int divisor) {
+	return (dividendo/(float)divisor);
 }
